requirements.txt
opencv-python
face_recognition
numpy
pillow
twilio 